+++ packages/engine/src/data/index.ts
@@ -3,4 +3,5 @@ export { LOGGING_ACTIONS } from './logging';
 export { HARVESTING_ACTIONS } from './harvesting';
 export { FISHING_ACTIONS } from './fishing';
 export { SCAVENGING_ACTIONS } from './scavenging';
+export { ASTROLOGY_ACTIONS } from './astrology';
 export { createNewPlayer } from './playerFactory';